通信処理を追加

+++ DestinationboardServer/Common/DBManager/SQLite/TablesBase/ActionMasterBase.cs
@@ -88,7 +88,7 @@ namespace DestinationboardServer.Common.DBManager.SQLite.TablesBase
 			}
 			set
 			{
-				if (!_CreateDate.Equals(value))
+				if (_CreateDate == null || !_CreateDate.Equals(value))
 				{
 					_CreateDate = value;
 					NotifyPropertyChanged("CreateDate");
@@ -140,7 +140,7 @@ namespace DestinationboardServer.Common.DBManager.SQLite.TablesBase
 			}
 			set
 			{
-				if (!_UpdateDate.Equals(value))
+				if (_UpdateDate == null || !_UpdateDate.Equals(value))
 				{
 					_UpdateDate = value;
 					NotifyPropertyChanged("UpdateDate");
@@ -252,10 +252,13 @@ namespace DestinationboardServer.Common.DBManager.SQLite.TablesBase
 		{
 			using (var db = new SQLiteDataContext())
 			{
-				var item = db.DbSet_ActionMaster.Single(x => x.ActionID.Equals(pk_item.ActionID));
-				item.Copy(update_item);
+				var item = db.DbSet_ActionMaster.SingleOrDefault(x => x.ActionID.Equals(pk_item.ActionID));
 
-				db.SaveChanges();
+				if (item != null)
+				{
+					item.Copy(update_item);
+					db.SaveChanges();
+				}
 			}
 		}
 		#endregion
@@ -269,9 +272,12 @@ namespace DestinationboardServer.Common.DBManager.SQLite.TablesBase
 		{
 			using (var db = new SQLiteDataContext())
 			{
-				var item = db.DbSet_ActionMaster.Single(x => x.ActionID.Equals(pk_item.ActionID));
-				db.DbSet_ActionMaster.Remove(item);
-				db.SaveChanges();
+				var item = db.DbSet_ActionMaster.SingleOrDefault(x => x.ActionID.Equals(pk_item.ActionID));
+				if (item != null)
+				{
+					db.DbSet_ActionMaster.Remove(item);
+					db.SaveChanges();
+				}
 			}
 		}
 		#endregion
