+++ DestinationboardServer/Common/DBManager/SQLite/TablesBase/ActionPlanTableBase.cs
@@ -336,10 +336,13 @@ namespace DestinationboardServer.Common.DBManager.SQLite.TablesBase
 		{
 			using (var db = new SQLiteDataContext())
 			{
-				var item = db.DbSet_ActionPlanTable.Single(x => x.StaffID.Equals(pk_item.StaffID));
-				item.Copy(update_item);
+				var item = db.DbSet_ActionPlanTable.SingleOrDefault(x => x.StaffID.Equals(pk_item.StaffID));
 
-				db.SaveChanges();
+				if (item != null)
+				{
+					item.Copy(update_item);
+					db.SaveChanges();
+				}
 			}
 		}
 		#endregion
@@ -353,9 +356,12 @@ namespace DestinationboardServer.Common.DBManager.SQLite.TablesBase
 		{
 			using (var db = new SQLiteDataContext())
 			{
-				var item = db.DbSet_ActionPlanTable.Single(x => x.StaffID.Equals(pk_item.StaffID));
-				db.DbSet_ActionPlanTable.Remove(item);
-				db.SaveChanges();
+				var item = db.DbSet_ActionPlanTable.SingleOrDefault(x => x.StaffID.Equals(pk_item.StaffID));
+				if (item != null)
+				{
+					db.DbSet_ActionPlanTable.Remove(item);
+					db.SaveChanges();
+				}
 			}
 		}
 		#endregion
